Organize GUI dialogs into submodule
