+++ src/gui/clipboard_dialog.rs
