+++ src/gui/notes_dialog.rs
