+++ src/gui/shell_cmd_dialog.rs
